fuck ahiihi

+++ command/micro-store/services/SellPostDetailService.js
@@ -1,6 +1,8 @@
 import { SellPostDetail } from '../models'
+import globalId from '../config/globalId'
+import { postRowCreatedPub, postRowDeletedPub, postRowUpdatedPub } from '../controllers/StorePubController'
 
-const SELLPOST_DETAIL_GLOBAL_ID = require('../config/globalId').SELLPOST_DETAIL_GLOBAL_ID
+const SELLPOST_DETAIL_GLOBAL_ID =  globalId.SELLPOST_DETAIL_GLOBAL_ID;
 
 export const getSellPostDetailLocalId = (id) => {
     if (id.length <= 3) return id;
@@ -43,6 +45,7 @@ export const updateSellPostDetail = (sellPostDetailId, updateInfo, next) => {
             if (updateInfo.products_order) sellPostDetail.productOrders = updateInfo.products_order;
             if (updateInfo.type) sellPostDetail.type = updateInfo.type;
             sellPostDetail.save(() => {
+                updateSellPostDetail(getPubBasicSellPostDetailInfo(sellPostDetail));
                 next(sellPostDetail)
             })
         }
@@ -50,17 +53,39 @@ export const updateSellPostDetail = (sellPostDetailId, updateInfo, next) => {
 };
 
 export const dellSellPostDetail = (sellPostDetailId, next) => {
-    SellPostDetail.findOneAndRemove({_id: getSellPostDetailLocalId(sellPostDetailId)}, () => {
-        next()
-    })
+    getSellPostDetail(sellPostDetailId, (sellPostDetail) => {
+       if (!sellPostDetail) next(null);
+        else {
+           SellPostDetail.findOneAndRemove({_id: getSellPostDetailLocalId(sellPostDetailId)}, () => {
+               postRowDeletedPub(sellPostDetailId, sellPostDetail.sellPostId);
+               next(sellPostDetail)
+           })
+       }
+    });
+
 };
 
 export const createSellPostDetail = (sellPostInfo, next) => {
     const sellPostDetail = new SellPostDetail({sellPostId: sellPostInfo.sellpostid, content: sellPostInfo.content,
                     line: sellPostInfo.numline, imageURLs: sellPostInfo.images, titlesOrder: sellPostInfo.titles_order,
-                    productOrders: sellPostInfo.products_order, type: sellPostInfo.type})
+                    productOrders: sellPostInfo.products_order, type: sellPostInfo.type});
     sellPostDetail.save(() => {
+        postRowCreatedPub(getPubBasicSellPostDetailInfo(sellPostDetail));
         next(sellPostInfo)
     })
+};
+
+export const getPubBasicSellPostDetailInfo = (sellPostDetail) => {
+    return {
+        sellPostId: sellPostDetail.sellPostId,
+        postrowId: getSellPostDetailGlobalId(sellPostDetail._id),
+        content: sellPostDetail.content,
+        numline: sellPostDetail.line,
+        images: sellPostDetail.imageURLs,
+        titles_order: sellPostDetail.titleOrder,
+        titles: sellPostDetail.titles,
+        products_order: sellPostDetail.productOrders,
+        type: sellPostDetail.type,
+    }
 };
